rename package and upgrade library
