+++ app/src/test/java/com/sun/mvp/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.sun.structure_android
+package com.sun.mvp
 
 import org.junit.Test
